users_service/requirements.txt
django
djangorestframework
# asgiref
# grpcio
# grpcio-tools
# protobuf
# python-dotenv
# pytz
# six
# sqlparse
psycopg2-binary